more Core/Aot unification
